requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
openai-whisper
numba==0.59.1
llvmlite==0.42.0
